Refactor common yaml package
