+++ shardingsphere-control-panel/shardingsphere-orchestration/shardingsphere-orchestration-core/shardingsphere-orchestration-core-common/src/test/java/org/apache/shardingsphere/orchestration/core/common/yaml/swapper/DataSourceConfigurationYamlSwapperTest.java
@@ -15,9 +15,10 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.orchestration.core.common.configuration;
+package org.apache.shardingsphere.orchestration.core.common.yaml.swapper;
 
 import org.apache.shardingsphere.infra.config.DataSourceConfiguration;
+import org.apache.shardingsphere.orchestration.core.common.yaml.config.YamlDataSourceConfiguration;
 import org.junit.Test;
 
 import static org.hamcrest.CoreMatchers.is;
